% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atsalibrary-package.R
\docType{package}
\name{atsalibrary-package}
\alias{atsalibrary-package}
\alias{atsalibrary}
\title{atsalibrary}
\description{
This package will load the needed packages and data files for the ATSA course material when students install from GitHub.
}
\section{Data}{

\itemize{
\item \link{chinook}
}

The rCAX R package includes a pull key so that you can query the CAX database. Should you
need to use your own key see the vignette on the API functions.
}

\keyword{package}
