% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bristol-bay-sockeye-data.R
\docType{data}
\name{bristol_bay_sockeye}
\alias{bristol_bay_sockeye}
\alias{bb_sockeye}
\title{Bristol Bay Sockeye Abundance and Forecasts}
\format{
Objects of class \code{"data.frame"}.
}
\source{
\href{https://github.com/DanOvando/salmon-forecast-paper}{DanOvando GitHub repo}

Daily escapement counts, Bristol Bay, Alaska, 1955-2017 \href{https://knb.ecoinformatics.org/view/doi\%3A10.5063\%2FHD7T35}{archived on KNB}
}
\usage{
data(bristol_bay_sockeye)
}
\description{
Bristol Bay sockeye abundance for 9 rivers for 4 age-groups. The data are from
Ovando et al 2021 Improving forecasts of sockeye salmon (Oncorhynchus nerka) with parametric and nonparametric models
DOI: 10.1139/cjfas-2021-0287. You’ll find a copy in the lab folder. The data file also has the covariates for year that
the smolts enter the ocean as used in Ovando et al. The dataset includes
climate covariates.
}
\details{
The data have the columns:
\itemize{
\item ret_yr The year the spawners return to the spawning grounds
\item ret The returns (number of fish in 1000s)
\item system The river name
\item age_group The age_group
\item forecast.adfw The forecast from AK Fish and Wildlife
\item forecast.fri The forecast from UW Fisheries Research Institute
\item env_* are some covariates at the year the age group entered the ocean
}

In the data file, the age group designation is “a.b” where “a” is number of years
in freshwater and “b” is number of years in the ocean. The age of the spawners in then a+b.
}
\examples{
data(bristol_bay_sockeye)
bb_sockeye \%>\% 
filter(system=="Kvichak") \%>\% 
  ggplot(aes(x=ret_yr, y=log(ret))) + 
    geom_line() + 
    ggtitle("log abundance by age group") +
    facet_wrap(~age_group)
}
\references{
Ovando et al 2021 Improving forecasts of sockeye salmon (Oncorhynchus nerka) with parametric and nonparametric modelDOI: 10.1139/cjfas-2021-0287
}
\keyword{datasets}
