% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esa-salmon-data.R
\docType{data}
\name{esa_salmon}
\alias{esa_salmon}
\alias{esa.salmon}
\alias{columbia.river}
\title{Annual spawner data from Endangered and Threatened PNW salmonids}
\format{
Objects of class \code{"data.frame"}.  Columns are species,
esu_dps, majorpopgroup, commonpopname, run, spawningyear, spawner, log.spawner
}
\source{
\href{https://www.streamnet.org/cap/current-hli/}{CAP} StreamNet Coordinated Assessments Partnership (CAP) standardized high-level indicators (HLIs) for Natural Origin Spawner Abundance (NOSA).
}
\usage{
data(esa_salmon)
}
\description{
The data set has yearly spawner counts for endangered and threatened ESU (Evolutionary Significant Units) and DPS (Distinct Population Segments) in the Washington and Oregon. Data were downloaded from  \href{https://www.streamnet.org/resources/exchange-tools/rest-api-documentation/}{Coordinated Assessments API}. Coordinated
Assessments data eXchange (CAX) is developed by the Coordinated Assessments Partnership (CAP).
}
\details{
There are two datasets included:  \code{esa.salmon} and \code{columbia.river}.  The Columbia River data set is a subset of the \code{esa.salmon} dataset that has all the ESUs and DPSs.
The dataset has the following columns
\itemize{
\item species: Chinook, Coho, Steelhead, Chum, Sockeye
\item esu_dps: name of the ESU
\item majorpopgroup: biological major group
\item commonpopname: common population name, generally a stream or river
\item run: run-timing
\item spawningyear: the year that the spawners were counted on the spawning grounds
\item spawner: total (natural-born and hatchery-born) spawners on the spawning ground. Generally some type of redd-count expansion or some other stream count of spawners. Redd = a gravel nest.
\item log.spawners: log of value
}
}
\examples{
data(esa.salmon)
df <- esa.salmon \%>\% subset(species == "Steelhead" & run == "Winter")
ggplot(df, aes(x=spawningyear, y=log.spawner, color=majorpopgroup)) + 
  geom_point(size=0.2) + 
  theme(strip.text.x = element_text(size = 2)) +
  theme(axis.text.x = element_text(size = 5, angle = 90)) +
  facet_wrap(~esapopname)
}
\references{
rCAX: https://zenodo.org/records/10214433
}
\keyword{datasets}
