% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourlyphyto.R
\docType{data}
\name{hourlyphyto}
\alias{hourlyphyto}
\title{Hourly phytoplankton}
\format{
Objects of class \code{"data.frame"} with column V1
}
\usage{
data(hourlyphyto)
}
\description{
no information about these data
}
\examples{
data(hourlyphyto)
}
\keyword{datasets}
