% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KvichakSockeye-data.R
\docType{data}
\name{KvichakSockeye}
\alias{KvichakSockeye}
\title{Kvichak River SR Data}
\format{
Object of class \code{"data.frame"}
\itemize{
  \item \strong{brood_year}  Brood year. 
  \item \strong{spawners} Spawners (count) in thousands 
  \item \strong{recruits} Recruits (count) in thousands
  \item \strong{pdo_summer_t2} Pacific Decadal Oscillation (PDO) from Apr-Sep of brood year $t+2$ 
  \item \strong{pdo_winter_ts} Pacific Decadal Oscillation (PDO) from Oct (brood year $t+2$) to Mar (brood year $t+3$)
}
}
\source{
\href{https://www.ramlegacy.org/}{RAM Legacy Stock Assessment Database}
}
\usage{
data(KvichakSockeye)
}
\description{
Spawner-recruit data for sockeye salmon from the Kvichak River.
}
\details{
Spawner-recruit data for sockeye salmon 
(_Oncorhynchus nerka_) from the Kvichak River
 in SW Alaska that span the years 1952-1989. The data come from a large 
 public database begun by Ransom Myers many years ago. The 
 database is now maintained as the RAM Legacy Stock Assessment Database.
}
\examples{
data(KvichakSockeye)
}
\references{
RAM Legacy Stock Assessment Database. 2018. Version 4.44-assessment-only. Released 2018-12-22. Retrieved from DOI:10.5281/zenodo.2542919.

Ricard, D., Minto, C., Jensen, O.P. and Baum, J.K. (2012) Evaluating the knowledge base and status of commercially exploited marine species with the RAM Legacy Stock Assessment Database. Fish and Fisheries 13 (4) 380-398. DOI: 10.1111/j.1467-2979.2011.00435.x

Spawner escapement estimates are originally from: 1997-2017 Appendix Table A12.
}
\seealso{
\link{sockeye}
}
\keyword{datasets}
