% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atsalibrary-package.R
\docType{package}
\name{atsalibrary-package}
\alias{atsalibrary-package}
\alias{atsalibrary}
\title{atsalibrary}
\description{
This package will load the needed packages and data files for the ATSA course material when students install from GitHub.
}
\section{Data}{

\itemize{
\item \link{chinook}
\item \link{ECNNO2}
\item \link{greeklandings}
\item \link{lakeWA}
\item \link{snotel}
\item \link{sockeye}
\item \link{neon_barc}
\item \link{MLCO2}
\item \link{NHTemp}
}
}

\keyword{package}
